RTPARSER -  <aiml> tag graph=".." is fixed.. not sure how long it was broken , pronoun resolution improved, heardSelfSay actualyl almost works

+++ trunk/RTParser/AIMLTagHandlers/learn.cs
@@ -159,7 +159,7 @@ namespace RTParser.AIMLTagHandlers
                             {
                                 try
                                 {
-                                    int successes = request.Loader.loadAIMLNode(templateNode, loaderOptions, request);
+                                    long successes = request.Loader.loadAIMLNode(templateNode, loaderOptions, request);
                                     return "" + successes;
                                 }
                                 finally
@@ -175,7 +175,7 @@ namespace RTParser.AIMLTagHandlers
                             {
                                 path = path ?? innerXML;
                                 loaderOptions.Loading0 = path;
-                                int forms = request.Loader.loadAIMLURI(path, loaderOptions);
+                                long forms = request.Loader.loadAIMLURI(path, loaderOptions);
                                 QueryHasSuceededN++;
                                 return "" + forms; // Succeed();
                             }
